Check if a tray exists and running before trying to remove it
During setup we want to force the installation of daemon service
so that it always run the crc binary which was used in the setup
but previously we were not checking if a service with the same name
exists or running, this patch adds that check before proceeding with
the kill

+++ pkg/crc/preflight/preflight_checks_tray_windows.go
@@ -35,9 +35,14 @@ func checkIfDaemonServiceInstalled() error {
 }
 
 func fixDaemonServiceInstalled() error {
-	// try to remove if a previous version exists
-	_ = service.Stop(constants.DaemonServiceName)
-	_ = service.Delete(constants.DaemonServiceName)
+	// only try to remove if a previous version exists
+	if service.IsInstalled(constants.DaemonServiceName) {
+		if service.IsRunning(constants.DaemonServiceName) {
+			_ = service.Stop(constants.DaemonServiceName)
+		}
+		_ = service.Delete(constants.DaemonServiceName)
+	}
+
 	// get executables path
 	binPath, err := goos.Executable()
 	if err != nil {
